+++ changelog.md
@@ -3,6 +3,7 @@ The latest version may not be released!
 See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-IRremote/commits/master
 
 ## 3.5.3
+- Separated enable flag of send and receive feedback LED. Inspired by PR#970 from luvaihassanali.
 - RP2040 support added.
 - Refactored IRTimer.hpp.
 - Refactored IR_SEND_PIN and IrSender.sendPin handling.